refactor(worklist): the cut moves to its own file
worklist.go crossed the 500-line cap when the resume landed in it. pageFrom
and keepCategory move to page.go: they are the cut, which is the concern the
cursor belongs to, and the rest of worklist.go is assembly and accounting.

Also repairs a doc comment that had already run two functions together on
main — worklistFrom's first line sat above waitingCustomers, so neither
said what it did.
